+++ sparse.hpp
@@ -6,11 +6,11 @@
 class SparseVector {
     private:
         std::vector<unsigned int> idxs;
-        std::vector<double> vals;
+        std::vector<long double> vals;
     
     public:
-        void AddEntry(unsigned int i, double val);
-        double Dot(std::vector<double>& v);
+        void AddEntry(unsigned int i, long double val);
+        long double Dot(std::vector<long double>& v);
 };
 
 #endif /* SPARSE_HPP */
